+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - export: new command. export generates a compressed file with the entire pot in it. It works only for single type pot
+- export: -D option to change export directory and -l option to change compression level
 - POT_CACHE: a place to cache pot images. It's a variable of pot.conf
 - import: new command. import create a new pot based on an image generated via export